+++ src/control_plane/shared_infrastructure/gotenberg/mod.rs
@@ -1,11 +1,21 @@
 pub(crate) use gotenberg_plan_error::GotenbergPlanError;
+pub(crate) use gotenberg_preparation_error::GotenbergPreparationError;
+pub(crate) use gotenberg_preparation_options::GotenbergPreparationOptions;
 pub(crate) use gotenberg_project_resources::GotenbergProjectResources;
 pub(crate) use gotenberg_shared_instance_plan::GotenbergSharedInstancePlan;
 pub(crate) use gotenberg_shared_instance_plan_options::GotenbergSharedInstancePlanOptions;
 pub(crate) use plan_gotenberg_project_resources::plan_gotenberg_project_resources;
+pub(crate) use prepare_gotenberg_shared_instances::prepare_gotenberg_shared_instances;
+pub(crate) use prepared_gotenberg_shared_instance::PreparedGotenbergSharedInstance;
+pub(crate) use reconcile_prepared_gotenberg_instance::reconcile_prepared_gotenberg_instance;
 
 mod gotenberg_plan_error;
+mod gotenberg_preparation_error;
+mod gotenberg_preparation_options;
 mod gotenberg_project_resources;
 mod gotenberg_shared_instance_plan;
 mod gotenberg_shared_instance_plan_options;
 mod plan_gotenberg_project_resources;
+mod prepare_gotenberg_shared_instances;
+mod prepared_gotenberg_shared_instance;
+mod reconcile_prepared_gotenberg_instance;